refactor(RepoWidget): Split data handling from widget
